refactor: collapse session history into session module and move ansi under ssh
