+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod ansi;
 pub mod config;
 #[macro_use]
 pub mod i18n;
@@ -6,6 +5,5 @@ pub mod gui;
 pub mod keychain;
 pub mod platform;
 pub mod session;
-pub mod session_history;
 pub mod ssh;
 pub mod terminal;
