+++ src/session/history.rs
